shopping_backend/requirements.txt
aioamqp==0.14.0
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-rabbitmq==1.3.1
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
dj-rest-auth==2.1.3
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.6.0
djangorestframework==3.12.2
hiredis==1.1.0
hyperlink==21.0.0
idna==3.1
importlib-metadata==1.7.0
incremental==17.5.0
msgpack==1.0.2
oauthlib==3.1.0
pamqp==2.3.0
pika==1.1.0
Pillow==8.1.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.1
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
urllib3==1.26.2
zipp==3.4.0
zope.interface==5.2.0
